Create XOX GAME
The name of the project is XOX. It is a game. The reason why we did this project is to kill our leisure time and for entertainment purpose.
This game is played on 3X3 grid by two players. the first player makes a move with 0. The player who has formed a horizontal,vertical or diagonal sequence of the three marks WINS. The game ends when the grid is filled completely.